ref: move code to utils
