+++ packages/design-system/src/components/cookieTable/index.tsx
@@ -25,15 +25,13 @@ import React, {
   useReducer,
 } from 'react';
 import { CookieTableData, getCookieKey } from '@ps-analysis-tool/common';
-import { saveAs } from 'file-saver';
-import classNames from 'classnames';
 
 /**
  * Internal dependencies.
  */
 import { Table, TableColumn, TableData, TableFilter, TableRow } from '../table';
 import { TableProvider } from '../table/useTable';
-import { generateCookieTableCSV } from '../table/utils';
+import { conditionalTableRowClassesHandler, exportCookies } from './utils';
 
 interface CookieTableProps {
   data: TableData[];
@@ -107,14 +105,6 @@ const CookieTable = forwardRef<
     [onRowClick, onRowContextMenu]
   );
 
-  const exportCookies = useCallback((rows: TableRow[]) => {
-    const _cookies = rows.map(({ originalData }) => originalData);
-    if (_cookies.length > 0 && 'parsedCookie' in _cookies[0]) {
-      const csvTextBlob = generateCookieTableCSV(_cookies as CookieTableData[]);
-      saveAs(csvTextBlob, 'Cookies Report.csv');
-    }
-  }, []);
-
   const getRowObjectKey = useCallback(
     (row: TableRow) =>
       getCookieKey(
@@ -141,62 +131,16 @@ const CookieTable = forwardRef<
     return key === null ? null : getCookieKey(key?.parsedCookie);
   }, [selectedFrameCookie]);
 
-  const conditionalTableRowClasses = useCallback(
-    // eslint-disable-next-line complexity
+  const _conditionalTableRowClassesHandler = useCallback(
     (row: TableRow, isRowFocused: boolean, rowIndex: number) => {
-      const rowKey = getRowObjectKey(row);
-      const isBlocked =
-        (row.originalData as CookieTableData)?.isBlocked ||
-        (((row.originalData as CookieTableData)?.isBlocked ?? false) &&
-          (row.originalData as CookieTableData)?.blockedReasons &&
-          (row.originalData as CookieTableData)?.blockedReasons?.length);
-      const isHighlighted = row.originalData?.highlighted;
-      const isDomainInAllowList = (row.originalData as CookieTableData)
-        ?.isDomainInAllowList;
-
-      const tableRowClassName = classNames(
-        isBlocked &&
-          (rowKey !== selectedKey
-            ? rowIndex % 2
-              ? 'dark:bg-flagged-row-even-dark bg-flagged-row-even-light'
-              : 'dark:bg-flagged-row-odd-dark bg-flagged-row-odd-light'
-            : isRowFocused
-            ? 'bg-gainsboro dark:bg-outer-space'
-            : 'bg-royal-blue text-white dark:bg-medium-persian-blue dark:text-chinese-silver'),
-        isDomainInAllowList &&
-          !isBlocked &&
-          (rowKey !== selectedKey
-            ? rowIndex % 2
-              ? 'dark:bg-jungle-green-dark bg-leaf-green-dark'
-              : 'dark:bg-jungle-green-light bg-leaf-green-light'
-            : isRowFocused
-            ? 'bg-gainsboro dark:bg-outer-space'
-            : 'bg-royal-blue text-white dark:bg-medium-persian-blue dark:text-chinese-silver'),
-        rowKey !== selectedKey &&
-          !isBlocked &&
-          !isDomainInAllowList &&
-          (rowIndex % 2
-            ? isHighlighted
-              ? 'bg-dirty-pink'
-              : 'bg-anti-flash-white dark:bg-charleston-green'
-            : isHighlighted
-            ? 'bg-dirty-pink text-dirty-red dark:text-dirty-red text-dirty-red'
-            : 'bg-white dark:bg-raisin-black'),
-        rowKey === selectedKey &&
-          !isBlocked &&
-          !isDomainInAllowList &&
-          (isRowFocused
-            ? isHighlighted
-              ? 'bg-dirty-red'
-              : 'bg-gainsboro dark:bg-outer-space'
-            : isHighlighted
-            ? 'bg-dirty-pink text-dirty-red'
-            : 'bg-royal-blue text-white dark:bg-medium-persian-blue dark:text-chinese-silver')
+      return conditionalTableRowClassesHandler(
+        row,
+        isRowFocused,
+        rowIndex,
+        selectedKey
       );
-
-      return tableRowClassName;
     },
-    [getRowObjectKey, selectedKey]
+    [selectedKey]
   );
 
   const hasVerticalBar = useCallback((row: TableRow) => {
@@ -223,7 +167,7 @@ const CookieTable = forwardRef<
         onRowClick={onRowClick}
         onRowContextMenu={onRowContextMenuHandler}
         getRowObjectKey={getRowObjectKey}
-        conditionalTableRowClasses={conditionalTableRowClasses}
+        conditionalTableRowClassesHandler={_conditionalTableRowClassesHandler}
         exportTableData={exportCookies}
         hasVerticalBar={hasVerticalBar}
       >
